requirements.txt
asgiref==3.7.2
certifi==2024.8.30
distlib==0.3.9
dj-database-url==2.3.0
Django==4.2.3
django-environ==0.11.2
djangorestframework==3.14.0
filelock==3.16.1
gunicorn==23.0.0
mysqlclient==2.2.0
packaging==24.1
pipenv==2024.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3
virtualenv==20.26.6